step 19 done

+++ rentACarSpring/src/main/java/com/etiya/rentACarSpring/businnes/request/InvoiceRequest/CreateInvoiceRequest.java
@@ -24,7 +24,7 @@ public class CreateInvoiceRequest {
 	@JsonFormat(pattern = "yyyy-MM-dd")
 	private Date createDate;
 	@NotNull
-	private Double totalPrice;
+	private int totalPrice;
 	@NotNull
 	@Min(1)
 	private int totalRentDay;
